General codebase cleanup and refactoring.
This commit includes the following changes:

- Ran ESLint to identify and fix linting issues, including unused variables.
- Manually reviewed the codebase for duplicate code and other cleanup opportunities.
- Refactored utility functions, consolidating them into `src/utils/utils.js`.
- Updated import paths for utility functions in various components.
- Removed unused variables and code sections identified by ESLint and manual review.
- Ensured all automated tests pass after the changes.
- Verified that the build process completes successfully.

The cleanup aims to improve code readability, maintainability, and reduce potential bugs.
