+++ package.json
@@ -21,7 +21,7 @@
   },
   "devDependencies": {
     "@axe-core/react": "^4.10.1",
-    "@eslint/js": "^9.25.0",
+    "@eslint/js": "^9.28.0",
     "@testing-library/jest-dom": "^6.6.3",
     "@testing-library/react": "^16.3.0",
     "@testing-library/user-event": "^14.6.1",
@@ -36,6 +36,7 @@
     "eslint-plugin-jsx-a11y": "^6.10.2",
     "eslint-plugin-react-hooks": "^5.2.0",
     "eslint-plugin-react-refresh": "^0.4.19",
+    "eslint-plugin-vitest": "^0.5.4",
     "globals": "^16.0.0",
     "happy-dom": "^17.5.6",
     "jest": "^29.7.0",
