+++ src/components/EarthquakeDetailModalComponent.test.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { render, fireEvent, act } from '@testing-library/react';
+import { render, act } from '@testing-library/react';
 import { MemoryRouter, Routes, Route } from 'react-router-dom';
 import { EarthquakeDataContext } from '../contexts/EarthquakeDataContext'; // Import context
