+++ src/components/ErrorBoundary.jsx
@@ -7,7 +7,7 @@ class ErrorBoundary extends React.Component {
     this.state = { hasError: false, error: null, errorInfo: null };
   }
 
-  static getDerivedStateFromError(error) {
+  static getDerivedStateFromError() { // error parameter removed
     // Update state so the next render will show the fallback UI.
     return { hasError: true };
   }
